Moved Pieces and board to a class file and header.
Initially thought of moving the board to its own file for neatness, but it's not going to
grow very much more than it is and you can't use the Pieces without the Board. So I
opted against that because  I would be forcing neatness just for the sake of it and I
would end up with a tiny class in a file on its own complicating the build.