Added a game loop file and made related changes.
Trying to get the game to accept user input via the console so that we can verify the functionality of
the board. Might need to put this on hold and make a full test environment before continuing. This is
all relevant to the final product, but I am fast approaching the point where I need to know for certain
that things are working before I put time and effort in the graphics.